Some basic refactoring
