+++ ts/src/index.ts
@@ -1,7 +1,9 @@
 import { Feature, FeatureCollection, GeometryObject } from 'geojson';
-import { Node, Relation, Way } from './osmobjs';
 import { purgeProps, RefElements } from './utils';
-import XmlParser from './xmlparser';
+import { XmlParser } from './xmlparser';
+import { Node } from './node';
+import { Way } from './way';
+import { Relation } from './relation';
 
 interface IOptions {
     completeFeature?: boolean;
